refactored test_admin_integration module into smaller unit test modules
